Validacion Tipo Producto Request

+++ SAI/app/Http/Controllers/Tipo_ProductoController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Tipo_Producto;
 use Laracast\Flash\Flash;
+use App\Http\Requests\TipoProductoRequest;
+use Auth;
+
 
 class Tipo_ProductoController extends Controller
 {
@@ -28,7 +31,15 @@ class Tipo_ProductoController extends Controller
      */
     public function create()
     {
-        return view ('admin.producto.tipo_producto.create');
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            return view ('admin.producto.tipo_producto.create');
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden registrar.')->error();
+            return redirect()->back();
+
+        }
+        
     }
 
     /**
@@ -37,13 +48,23 @@ class Tipo_ProductoController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(TipoProductoRequest $request)
     {
-        $tipo_producto = new Tipo_Producto($request->all());
-        $tipo_producto->save();
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            $tipo_producto = new Tipo_Producto($request->all());
+            $tipo_producto->save();
 
-        flash("Registro del tipo producto '' " .$request->tip_tipo . " '' exitosamente.")->success();
-        return redirect()->route('tipo_producto.index');
+            flash("Registro del tipo producto '' " .$request->tip_tipo . " '' exitosamente.")->success();
+            return redirect()->route('tipo_producto.index');
+
+
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden registrar.')->error();
+            return redirect()->back();
+
+        }
+        
     }
 
     /**
@@ -65,9 +86,25 @@ class Tipo_ProductoController extends Controller
      */
     public function edit($id)
     {
-        $tipo_producto = Tipo_Producto::find($id);
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            $tipo_producto = Tipo_Producto::find($id);
+            if ($tipo_producto !== null) {
+                return view('admin.producto.tipo_producto.edit')->with(compact('tipo_producto'));
 
-        return view('admin.producto.tipo_producto.edit')->with(compact('tipo_producto'));
+
+
+            }else{  
+                flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+                return redirect()->route('producto_articulo.index');
+            }
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden modificar.')->error();
+            return redirect()->back();
+
+        }
+        
+
     }
 
     /**
@@ -77,15 +114,31 @@ class Tipo_ProductoController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(TipoProductoRequest $request, $id)
     {
-        $tipo_producto = Tipo_Producto::find($id);
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            $tipo_producto = Tipo_Producto::find($id);
+            if ($tipo_producto !== null) {
+                $tipo_producto->tip_tipo = $request->tip_tipo;
+                $tipo_producto->save();
 
-        $tipo_producto->tip_tipo = $request->tip_tipo;
-        $tipo_producto->save();
+                flash("Edición del tipo de producto '' " .$tipo_producto->tip_tipo . " '' exitosamente.")->success();
+                return redirect()->route('tipo_producto.index');
 
-        flash("Edición del tipo de producto '' " .$tipo_producto->tip_tipo . " '' exitosamente.")->success();
-        return redirect()->route('tipo_producto.index');
+
+            }else{  
+                flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+                return redirect()->route('producto_articulo.index');
+            }
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden modificar.')->error();
+            return redirect()->back();
+
+        }
+        
+
+        
     }
 
     /**
@@ -96,10 +149,24 @@ class Tipo_ProductoController extends Controller
      */
     public function destroy($id)
     {
-        $tipo_producto = Tipo_Producto::find($id);
-        $tipo_producto->delete();
+        if (Auth::user()->rol->rol_rol === 'Administrador'){
+            $tipo_producto = Tipo_Producto::find($id);
+            if ($tipo_producto !== null) {
+                $tipo_producto->delete();
 
-        flash("Eliminación del tipo de producto '' " .$tipo_producto->tip_tipo . " '' exitosamente.")->success();
-        return redirect()->route('tipo_producto.index');
+                flash("Eliminación del tipo de producto '' " .$tipo_producto->tip_tipo . " '' exitosamente.")->success();
+                return redirect()->route('tipo_producto.index');
+            }else{  
+                flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+                return redirect()->route('producto_articulo.index');
+            }
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador"  puede eliminar.')->error();
+            return redirect()->back();
+
+        }
+        
+        
     }
 }
